# holto/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HOLTO travel disruption companion API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  createdAt: string;
  starterPackEmail?: string | null;
}

export interface RegisterBody {
  email: string;
  /** @minLength 8 */
  password: string;
  /** @minLength 1 */
  name: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface StarterPackBody {
  email: string;
}

export type DisruptionType = typeof DisruptionType[keyof typeof DisruptionType];


export const DisruptionType = {
  delay: 'delay',
  cancellation: 'cancellation',
  missed_connection: 'missed_connection',
  denied_boarding: 'denied_boarding',
} as const;

export interface ChecklistItem {
  id: string;
  text: string;
  done: boolean;
  category?: string;
}

export interface ActionItem {
  order: number;
  title: string;
  description: string;
}

export type ProactiveActionUrgency = typeof ProactiveActionUrgency[keyof typeof ProactiveActionUrgency];


export const ProactiveActionUrgency = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface ProactiveAction {
  title: string;
  description: string;
  urgency: ProactiveActionUrgency;
  why: string;
}

export interface Disruption {
  id: number;
  userId: number;
  airline: string;
  flightNumber: string;
  origin: string;
  destination: string;
  scheduledAt: string;
  disruptionType: DisruptionType;
  details: string;
  rights?: string | null;
  actions?: ActionItem[] | null;
  checklist?: ChecklistItem[] | null;
  companionMessage?: string | null;
  proactiveHint?: string | null;
  proactiveAction?: ProactiveAction | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateDisruptionBody {
  airline: string;
  flightNumber: string;
  origin: string;
  destination: string;
  scheduledAt: string;
  disruptionType: DisruptionType;
  details: string;
}

export interface UpdateDisruptionBody {
  checklist?: ChecklistItem[];
}

export interface AnalyzeBody {
  airline: string;
  flightNumber: string;
  origin: string;
  destination: string;
  scheduledAt: string;
  disruptionType: DisruptionType;
  details: string;
}

export interface CompanionAnalysis {
  rights: string;
  actions: ActionItem[];
  checklist: ChecklistItem[];
  companionMessage: string;
  proactiveHint?: string | null;
  proactiveAction?: ProactiveAction | null;
}

export type FlightStatusStatus = typeof FlightStatusStatus[keyof typeof FlightStatusStatus];


export const FlightStatusStatus = {
  scheduled: 'scheduled',
  active: 'active',
  landed: 'landed',
  cancelled: 'cancelled',
  incident: 'incident',
  diverted: 'diverted',
  unknown: 'unknown',
} as const;

export interface FlightStatus {
  flightNumber: string;
  airlineIata?: string | null;
  status: FlightStatusStatus;
  depAirport?: string | null;
  arrAirport?: string | null;
  scheduledDep?: string | null;
  estimatedDep?: string | null;
  scheduledArr?: string | null;
  estimatedArr?: string | null;
  depDelay?: number | null;
  arrDelay?: number | null;
  depGate?: string | null;
  depTerminal?: string | null;
  arrTerminal?: string | null;
  companionMessage?: string | null;
  checkedAt: string;
}

export interface MonitoredFlight {
  id: number;
  flightNumber: string;
  destination: string;
  lastStatus?: string | null;
  lastCheckedAt?: string | null;
  createdAt: string;
}

export interface AddMonitorBody {
  flightNumber: string;
  destination: string;
}

export interface NearbyPlace {
  name: string;
  vicinity: string;
  rating?: number | null;
  openNow?: boolean | null;
  placeId?: string | null;
  lat?: number | null;
  lng?: number | null;
}

export interface AskBody {
  question: string;
  lat?: number | null;
  lng?: number | null;
}

export interface AskResponse {
  answer: string;
  places: NearbyPlace[];
}

export type GetFlightStatusParams = {
/**
 * IATA flight number (e.g. BA245)
 */
flightNumber: string;
};